% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Defense}
\alias{Defense}
\title{Weekly player Defense statistics (2012-2014)}
\format{A dataframe of 2448 rows and 15 variables:
\itemize{
\item Name: Player name
\item Team: Player team
\item G: Game or week
\item Solo: Solo tackles
\item Ast: Assited tackles
\item Tot: Total tackles (Solo + Ast)
\item Sacks: Number of Sacks
\item Yds: Yards lost from sacks
\item Int: Interceptions
\item Yds1: Yds from interceptions
\item TD: Touchdowns from interceptions
\item FrcdFum: Forced Fumbles
\item PassDef: Passes Deflected
\item Week: Week of the season
\item Year: Year or season
}}
\usage{
data(Defense)
}
\description{
A dataset containing Defensive statistics for NFL players from the 2012 through 2014 seasons
}
\keyword{data}

