% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Retrieve stats data}
\usage{
GetData(URI)
}
\arguments{
\item{URI}{string used to retrieve the relevant stats table}
}
\value{
a dataframe with the player statistics
}
\description{
Uses the inputs to create a URI and then scrapes
the Yahoo page to return desired the stats table.
}

