% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetGameID.R
\name{GetGameID}
\alias{GetGameID}
\title{Function to pull the game meta data}
\usage{
GetGameID(sig)
}
\arguments{
\item{sig}{output from the GetSig() function}
}
\value{
Game ID number as character
}
\description{
The function makes a call to the Yahoo API and
returns the game data. Games refer to the seasonal
fantasy footbal competition and are the highest
level in the API heirarchy. Games should not be confused
with an weekly NFL football game.
}

