% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetLeagueData.R
\name{GetLeagueData}
\alias{GetLeagueData}
\title{Pull the league data for your league}
\usage{
GetLeagueData(collection, resource, sig)
}
\arguments{
\item{collection}{one of the collections from the
Yahoo API.  Possible values are "players",
"League", ""Teams", "Users",}

\item{resource}{resources are a subcategory of
collections.}

\item{sig}{output from the GetSig() function}
}
\value{
a dataframe frame with the results of the API call
}
\description{
Scrapes the Yahoo page from the given URI and returns
desited the stats table.
}

