% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetSig.R
\name{GetSig}
\alias{GetSig}
\title{Get authenticated to use the Yahoo API}
\usage{
GetSig(key, secret)
}
\arguments{
\item{key}{string, provided by the Yahoo Developer network
when creating the application.}

\item{secret}{string, provided by the Yahoo Developer network
when creating the application.}
}
\value{
a signature object used to authenticate API calls
}
\description{
Authenticates the user for Yahoo API.  Requires the user
to supply a key and secret after setting up their app
on the Yahoo Developer Network
}

