% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetStats.R
\name{GetStats}
\alias{GetStats}
\title{This function allows you to get player stats from Yahoo sports}
\usage{
GetStats(category, year, week)
}
\arguments{
\item{category}{Quoted string. One of 7 categories corresponding to the
types of data available.  Must be one of:
\itemize{
\item "Defense"
\item "Kicking"
\item "Passing"
\item "Punting"
\item "Receiving"
\item "Returns"
\item "Rushing"
}}

\item{year}{Four digit year starting in 2000}

\item{week}{Week of the season starting with 1 and ending
with 17.  Note: Post season data cannot be pulled with this
function.}
}
\value{
A dataframe with the desired stats
}
\description{
This function allows you to get player stats from Yahoo sports
}
\examples{
GetStats("Passing",2014,1)
}

