% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetTeamData.R
\name{GetTeamData}
\alias{GetTeamData}
\title{Function to return the team data}
\usage{
GetTeamData(GameID, LeagueID, team = 1, DataType = "stats", sig)
}
\arguments{
\item{GameID}{Character value.  The number assigned
to the annual fantasy footbal compettion. Games should
not be confused with a weekly NFL football game. See
alse the GetGameID() function.}

\item{LeagueID}{Character value. The ID number
representing the specific fantasy football league.}

\item{team}{Character value. The team ID, possibley values
are based on the number of teams in the league: typically
ranging from 1-10.}

\item{DataType}{Character value. The type of data desired
for the team. Possible values are stats, standings, or
matchups.}

\item{sig}{See also the GetSig() function}
}
\value{
List. Fiels will vary by the datatype used.
}
\description{
Returns data on the team
}

