% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Kicking}
\alias{Kicking}
\title{Weekly player Kicking statistics (2012-2014)}
\format{A dataframe of 2448 rows and 17 variables:
\itemize{
\item Name: Player name
\item Team: Player team
\item G: Game or week of the season
\item 0-19: Field goals attempted from 0 to 19 yards out,  The first digit is
the completed and second is the number attempted. e.g. 2-3
\item 20-29:
\item 30-39:
\item 40-49
\item 50+:
\item FGM: Field goals made
\item FGA: Field goals attempted
\item Pct: Percent field goals made
\item Lng: Long, the longest attempt completed
\item XPM: Extra points made
\item XPA: Extra points attempted
\item Pct.1:  Percent field goals made
\item Week: Week of the season
\item Year: Year or season
}}
\usage{
data(Kicking)
}
\description{
A dataset containing Kicking statistics for NFL players from the 2012 through 2014 seasons
}
\keyword{data}

