% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Passing}
\alias{Passing}
\title{Weekly player passing statistics (2012-2014)}
\format{A dataframe of 2448 rows and 17 variables:
\itemize{
\item Name: Player name
\item Team: Player team
\item QBRat: Quarterback rating
\item Comp: Completed passes
\item Att: Attempted passes
\item Yds: Yards from completed passes
\item Y/A: Yards after catch
\item Lng: Longest pass
\item Int: Number of interceptions
\item TD: Number of touchdowns
\item Sack: Number of times the quarteback was sacked
\item SackYds: Yard lost after sacks
\item Fum: Fumble, both lost and recovered
\item FumL: Fumbles lost to the defense
\item Week: Week of the season
\item Year: year or season
}}
\usage{
data(Passing)
}
\description{
A dataset containing passing statistics for NFL players from the 2012 through 2014 seasons
}
\keyword{data}

