% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{PlayerStats}
\alias{PlayerStats}
\title{2014 player statistics}
\format{A list of 6 dataframes each of different categories:}
\usage{
data(PlayerStats)
}
\description{
A dataset containing 6 different categories of statistics.
categories are as follows:
\itemize{
\item Passing: Stats such as QB Rating, Completions,
Attempts, and Touchdowns
\item Rushing: Stats such as Touchdowns, Fumbles,
Yards, and Touchdowns
\item Kicking: Stats such as completes by yardage bin and
completion percent
\item Returns: Stats such as number of kick returns, yards,
 and the length of the longest kick.
\item Punting Status such as number of touchbacks, blocks,
and average yards
\item Defense: Stats such as Sacks, interceptions, forced
fumbles, and touchdowns.
}
}
\keyword{data}

