% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Punting}
\alias{Punting}
\title{Weekly player Punting statistics (2012-2014)}
\format{A dataframe of 2448 rows by 17 variables:
\itemize{
\item Name: Player name
\item Team: Player team
\item G: Game or week of the season
\item Punt: Number of punts
\item Yds: Total yards
\item Avg: Average yards per punt
\item Long: Longest punt during the week
\item In20: Number of punts falling between the 20 yard line and the endzone
\item FC: Fair catches
\item TB: Touchback
\item Blk: Blocked
\item Week: Week of the season
\item Year: Year or season
}}
\usage{
data(Punting)
}
\description{
A dataset containing Punting statistics for NFL players from the 2012 through 2014 seasons
}
\keyword{data}

