% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Returns}
\alias{Returns}
\title{Weekly player Punting statistics (2012-2014)}
\format{A dataframe of 2448 rows by 17 variables:
\itemize{
\item Name: Player name
\item Team: Player team
\item G: Game or week of the season
\item KR: Number Kick off returns
\item Yds: Total yards returned from kickoff returns
\item Avg: Average yards from kickoff returns
\item Long: Longest kickoff return
\item TD: Number of Touchdowns
\item PR: Number of punt returns
\item Yds.a: Total yards from punt returns
\item Avg.1: Average yards from punt returns
\item Long.1: Longest punt return
\item TD.1: Touchdown from punt returns
\item Week: Week of the season
\item Year: Year or season
}}
\usage{
data(Punting)
}
\description{
A dataset containing Return statistics for NFL players from the 2012 through 2014 seasons
}
\keyword{data}

