% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Rushing}
\alias{Rushing}
\title{Weekly player Rushing statistics (2012-2014)}
\format{A dataframe of 2448 rows by 17 variables:
\itemize{
\item Name: Player name
\item Team: Player team
\item G: Game or week of the season
\item Att: Attempts
\item Yds: Total rushing yards
\item Y/A: Yards per attempt
\item Lng: Longest Rush
\item TD: Number of Touchdowns
\item Fum: Fumbles
\item FumL: Fumbles lost
\item Week: Week of the season
\item Year: Year or season
}}
\usage{
data(Punting)
}
\description{
A dataset containing Rushing statistics for NFL players from the 2012 through 2014 seasons
}
\keyword{data}

