% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ScoreKicking.R
\name{ScoreKicking}
\alias{ScoreKicking}
\title{Score the kicking stats}
\usage{
ScoreKicking(x, FieldGoal19 = 3, FieldGoal29 = 3, FieldGoal39 = 3,
  FieldGoal49 = 4, FieldGoal50 = 5, ExtraPoint = 1)
}
\arguments{
\item{x}{dataframe of player statistics data. Should be
of the format produced by the GetStats() function.}

\item{FieldGoal19}{Points awarded for field goals with
a range of 19 yards or less.}

\item{FieldGoal29}{Points awarded for field goals with
range greater than 19 yards and a far as 29 yards}

\item{FieldGoal39}{Points awarded for field goals with
range greater than 29 yards and a far as 39 yards}

\item{FieldGoal49}{Points awarded for field goals with
range greater than 39 yards and a far as 49 yards}

\item{FieldGoal50}{Points awarded for field goals with
range greater than 50 yards}

\item{ExtraPoint}{Points awarded for extra points}
}
\value{
1 by nrow(x) matrix with the total score for each row
}
\description{
Applies scoring rules to the player kicking statistics.
Default values are set to the standard scoring values.
}

