% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ScoreReceiving.R
\name{ScoreReceiving}
\alias{ScoreReceiving}
\title{Score the receiving stats}
\usage{
ScoreReceiving(x, ReceivingYds = c(10, 1), ReceivingTD = 4)
}
\arguments{
\item{x}{dataframe of player statistics data. Should be
of the format produced by the GetStats() function.}

\item{ReceivingYds}{Vector with the yardage increment and
point value. As an example, a player is awarded 1 point for
every 10 yards, then the score rule is of the format c(10,1)}

\item{ReceivingTD}{Points awarded for receiving touchdowns}
}
\value{
1 by nrow(x) matrix with the total score for each row
}
\description{
Applies scoring rules to the player statistics. Default
values are set to the standard scoring values.
}

