% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ScoreRushing.R
\name{ScoreRushing}
\alias{ScoreRushing}
\title{Score the rushing stats}
\usage{
ScoreRushing(x, RushingYds = c(10, 1), RushingTD = 6, FumbleLost = -2)
}
\arguments{
\item{x}{dataframe of player statistics data. Should be
of the format produced by the GetStats() function.}

\item{RushingYds}{Vector with the yardage increment and
point value. As an example, a player is awarded 1 point for
every 10 yards, then the score rule is of the format c(10,1)}

\item{RushingTD}{Number of points awarded for rushing touchdowns}

\item{FumbleLost}{Number of point awarded for each fumble lost.
Since this function is for offensive stats, a fumble lost would occur
when the defense recovers the fumble.}
}
\value{
1 by nrow(x) matrix with the total score for each row
}
\description{
Applies scoring rules to the player statistics
}

