% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GenerateURI.R
\name{GenerateURI}
\alias{GenerateURI}
\title{This function generates the URI needed to pull the data}
\usage{
GenerateURI(category, year, week)
}
\arguments{
\item{category}{One of 7 categories corresponding to the
types of data available.  Must be one of the following:
\itemize{
\item Passing
\item Rushing
\item Receiving
\item Kicking
\item Returns
\item Punting
\item Defense
}}

\item{year}{Four digit year starting in 2000}

\item{week}{Vector of Weeks of the season starting
with 1 and ending with 17.}
}
\value{
A string with the URI needed to query the Yahoo
 sports statistics
}
\description{
This function generates the URI needed to pull the data
}

