% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{Retrieve stats data}
\usage{
GetData(URI)
}
\arguments{
\item{URI}{string value. Properly formatted URI string.
When sent to a browser, this URI would show Yahoo's
NFL player stats page.}
}
\value{
a dataframe with the player statistics
}
\description{
Reads the HTML table with player statistics based on
}

