% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetSig.R
\name{GetSig}
\alias{GetSig}
\title{Sets up the OAuth signature credential with the Yahoo API}
\usage{
GetSig(key, secret)
}
\arguments{
\item{key}{String value. Provided by the Yahoo Developer Network
when creating the application.}

\item{secret}{String value. Provided by the Yahoo Developer Network
when creating the application.}
}
\value{
This is called for its side effect. Creates a signature
object used to authenticate API calls when using the GetLeagueData
function.
}
\description{
Authenticates the user for Yahoo API.  Requires the user
to supply a key and secret after setting up their app
on the Yahoo Developer Network
}

