% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/GetStats.R
\name{GetStats}
\alias{GetStats}
\title{To get player stats from Yahoo sports}
\usage{
GetStats(category, year, week)
}
\arguments{
\item{category}{once of 7 categories corresponding to the
types of data available.  Must be one of Passing",
"Rushing, Receiving, Kicking, Returns, Punting",
"or Defense"}

\item{year}{four digit year starting in 2000}

\item{week}{week of the season starting with 1 and ending
with 17.  Post season data cannot be pulled with this
function.}
}
\value{
returns a dataframe with the desired stats
}
\description{
To get player stats from Yahoo sports
}

