% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Touchdown.R
\docType{data}
\name{Passing}
\alias{Passing}
\title{2014 player passing statistics}
\usage{
data(Passing)
}
\description{
A dataset containing passing statistics for NFL players from 2014
\itemize{
\item Name: Player name
\item Team: Player team
\item QBRat: Quarterback rating
\item Comp: Completed pases
\item Att: Attempted pases
\item Yds: Yards from completed passes
\item Y/A: Yards after cathc
\item Lng: Longest pass
\item Int: Number of interceptions
\item TD: Number of touchdowns
\item Sack: Number of times the quarteback was sacked
\item SackYds: Yard lost after sacks
\item Fum: Fumble, both lost and recovered
\item FumL: Fumbles lost to the defense
\item Week: Week of the season
}
}
\keyword{data}

